Extract revenue logic from article_form_controller (#1839)
Splits the 640-line article_form_controller.js by lifting the
revenue-ratio cluster (selection, calculation, summary, validation) into a
new article_revenue_controller.js sibling. The orchestrator keeps
autosave, draft recovery, save-status, leave-warning, editor modes
(focus/preview/settings), and currency handling.

Cross-controller coordination is one-way (revenue -> form): the revenue
controller dispatches `article-revenue:queue-autosave`, which the
orchestrator listens for on the shared editor root and forwards to its
debounced autosave. No Stimulus outlets (the codebase has none and the
outlet accessor throws on a missing match); a plain CustomEvent is
simpler and safer for a single-method dependency.

Audit findings (issue #1839):
- The controller root lives in three entry points (new.html.erb,
  _edit_form.html.erb, + shared partials), not two as the issue stated.
  This does a focused 2-way split rather than the issue's 4-way outlet
  split to keep the high-risk autosave/preview/conflict flows intact.
- `article-form#touchDirty` is bound in 3 views but the method never
  existed -- a silent no-op. Preserved verbatim with an empty stub +
  comment so behavior is unchanged; fixing it is a separate concern.

Behavior-preserving: all revenue targets/actions repointed to
article-revenue; currency (calPriceUsd, priceUsd target) and the
touchDirty no-op stay on article-form. articlePublished is supplied to
both controllers since the currency guard and the revenue guard both
read it.

Verified: prettier lint-check, zeitwerk:check, stimulus:manifest:update
(runs clean; index.js is manually maintained to avoid generator
reformatting churn), and test/system/article_paywall_test.rb +
test/controllers/articles_controller_test.rb +
test/controllers/admin/articles_controller_test.rb all pass.

Closes #1839.
