+++ app/views/article_references/_form.html.erb
@@ -35,10 +35,10 @@
         <%= form.number_field :revenue_ratio, 
           required: true, 
           disabled: article.published_at.present?,
-          data: { 
-            article_form_target: "articleReferenceRevenueRatio",
-            action: "article-form#calReferenceRatio blur->article-form#formatReferenceRatio article-form#touchDirty"
-          }, 
+          data: {
+            article_revenue_target: "articleReferenceRevenueRatio",
+            action: "article-revenue#calReferenceRatio blur->article-revenue#formatReferenceRatio article-form#touchDirty"
+          },
           step: "0.01",
           min: "0.01",
           max: "0.5",
@@ -51,7 +51,7 @@
   <% if article.published_at.blank? %>
     <div class="flex justify-end">
       <div 
-        data-action="click->nested-form#remove click->article-form#calReferenceRatio click->article-form#touchDirty"
+        data-action="click->nested-form#remove click->article-revenue#calReferenceRatio click->article-form#touchDirty"
         class="cursor-pointer px-4 py-2 font-bold text-center border border-dashed border-red-500 text-red-500 rounded-full">
         <%= t('articles.remove_reference') %>
       </div>
